+++ ta4j-core/src/test/java/org/ta4j/core/indicators/helpers/TradeCountIndicatorTest.java
@@ -1,7 +1,7 @@
 /**
  * The MIT License (MIT)
  *
- * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2019 Ta4j Organization & respective
+ * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2021 Ta4j Organization & respective
  * authors (see AUTHORS)
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -23,6 +23,10 @@
  */
 package org.ta4j.core.indicators.helpers;
 
+import static org.junit.Assert.assertEquals;
+
+import java.util.function.Function;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.ta4j.core.BarSeries;
@@ -30,10 +34,6 @@ import org.ta4j.core.Indicator;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
-
-import java.util.function.Function;
-
-import static org.junit.Assert.assertEquals;
 
 public class TradeCountIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
     private TradeCountIndicator tradeIndicator;
@@ -53,7 +53,7 @@ public class TradeCountIndicatorTest extends AbstractIndicatorTest<Indicator<Num
     @Test
     public void indicatorShouldRetrieveBarTrade() {
         for (int i = 0; i < 10; i++) {
-            assertEquals((int) tradeIndicator.getValue(i), barSeries.getBar(i).getTrades());
+            assertEquals((long) tradeIndicator.getValue(i), barSeries.getBar(i).getTrades());
         }
     }
 }
